implemented clic peripheral

+++ src/interrupt.rs
@@ -66,5 +66,6 @@ pub unsafe trait InterruptNumber: Copy {
     /// Return the interrupt number associated with this variant.
     ///
     /// See trait documentation for safety requirements.
-    fn number(self) -> u32;
+    fn number(self) -> usize;
 }
+
